Redo level counting stuff
